Chat room messages CRUD

+++ base/views.py
@@ -1,6 +1,12 @@
 from django.shortcuts import render, redirect
-from .models import Room, Topic
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic, Message
 from django.db.models import Q
+from django.contrib.auth.models import User 
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
 from .forms import RoomForm
 # rooms=[
 #     {'id': 1, 'name':'Lets learn python'},
@@ -9,6 +15,46 @@ from .forms import RoomForm
 
 # ]
 # Create your views here.
+def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method=='POST':
+        username = request.POST.get('Username')
+        password = request.POST.get('Password')
+        try:
+            user=User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exists')
+            
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or password does not exitsts')
+
+    context = {'page':page}
+    return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form= UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during the registration')
+    return render(request,'base/login_register.html', {'form':form})
+
 def home(request):
     q=request.GET.get('q') if request.GET.get('q')!=None else ''
     rooms=Room.objects.filter(
@@ -23,9 +69,22 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_messages=room.message_set.all().order_by('-created')
+    participants=room.participants.all()
+
+    if request.method=='POST':
+        message=Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room':room,'room_messages':room_messages, 'participants':participants}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,9 +95,13 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user!= room.host:
+        return HttpResponse("You cannot update other rooms!")
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -48,8 +111,11 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!= room.host:
+        return HttpResponse("You cannot delete other rooms!")
     if request.method=='POST':
         room.delete()
         return redirect('home')
